apps/backend/pyproject.toml
[project]
name = "localforge-backend"
version = "0.1.0"
description = "FastAPI backend for LocalForge"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]==0.115.5",
    "httpx>=0.28.1",
    "dnspython==2.8.0",
    "openpyxl==3.1.5",
    "pandas==2.2.2",
    "pillow==10.4.0",
    "pypdf==4.3.1",
    "python-multipart==0.0.9",
    "uvicorn[standard]==0.30.6",
]

[dependency-groups]
dev = [
    "python-docx==1.1.2",
    "pytest==8.3.2",
    "requests==2.32.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
